+++ CHANGELOG.md
@@ -5,4 +5,5 @@ All notable changes to `laravel-disallowlister` will be documented in this file
 ## 0.1.0 2021-08-23
 - Created accentinteractive/laravel-disallowlister
 - Added facade `Disallowlister`
+- Added validation rule `disallowlister`
 - Added publishable config file `disallowlister.php` to contain an default array of disallowed strings.